+++ src/Sentry.DiagnosticSource/Internals/DiagnosticSource/SentryEFCoreListener.cs
@@ -57,23 +57,46 @@ namespace Sentry.Internals.DiagnosticSource
 
         private ISpan? AddSpan(SentryEFSpanType type, string operation, string? description)
         {
-            if (_hub.GetSpan()?.StartChild(operation, description) is { } span &&
-                GetSpanBucket(type) is { } asyncLocalSpan)
+            ISpan? span = null;
+            _hub.ConfigureScope(scope =>
             {
-                asyncLocalSpan.Value = new WeakReference<ISpan>(span);
-                return span;
-            }
-            return null;
+                if (scope.Transaction?.IsSampled != true)
+                {
+                    return;
+                }
+
+                if (scope.GetSpan()?.StartChild(operation, description) is not { } startedChild)
+                {
+                    return;
+                }
+
+                if (GetSpanBucket(type) is not { } asyncLocalSpan)
+                {
+                    return;
+                }
+
+                asyncLocalSpan.Value = new WeakReference<ISpan>(startedChild);
+                span = startedChild;
+            });
+            return span;
         }
 
         private ISpan? TakeSpan(SentryEFSpanType type)
         {
-            if (GetSpanBucket(type)?.Value is { } reference && reference.TryGetTarget(out var span))
+            ISpan? span = null;
+            _hub.ConfigureScope(scope =>
             {
-                return span;
-            }
-            _options.DiagnosticLogger?.LogWarning("Trying to close a span that was already garbage collected. {0}", type);
-            return null;
+                if (scope.Transaction?.IsSampled == true)
+                {
+                    if (GetSpanBucket(type)?.Value is { } reference &&
+                        reference.TryGetTarget(out var startedSpan))
+                    {
+                        span = startedSpan;
+                    }
+                    _options.LogWarning("Trying to close a span that was already garbage collected. {0}", type);
+                }
+            });
+            return span;
         }
 
         private AsyncLocal<WeakReference<ISpan>>? GetSpanBucket(SentryEFSpanType type)
@@ -105,7 +128,7 @@ namespace Sentry.Internals.DiagnosticSource
 
                 //Connection Span
                 //A transaction may or may not show a connection with it.
-                if (_logConnectionEnabled && value.Key == EFConnectionOpening)
+                else if (_logConnectionEnabled && value.Key == EFConnectionOpening)
                 {
                     AddSpan(SentryEFSpanType.Connection, "db.connection", null);
                 }
@@ -133,7 +156,7 @@ namespace Sentry.Internals.DiagnosticSource
             }
             catch (Exception ex)
             {
-                _options.DiagnosticLogger?.LogError("Failed to intercept EF Core event.", ex);
+                _options.LogError("Failed to intercept EF Core event.", ex);
             }
         }