+++ src/Sentry/TransactionTracer.cs
@@ -222,19 +222,14 @@ namespace Sentry
         /// <inheritdoc />
         public void Finish()
         {
-            try
-            {
-                Status ??= SpanStatus.UnknownError;
-                EndTimestamp = DateTimeOffset.UtcNow;
+            Status ??= SpanStatus.UnknownError;
+            EndTimestamp = DateTimeOffset.UtcNow;
 
-                // Client decides whether to discard this transaction based on sampling
-                _hub.CaptureTransaction(new Transaction(this));
-            }
-            finally
-            {
-                // Clear the transaction from the scope
-                _hub.ConfigureScope(scope => scope.ResetTransaction(this));
-            }
+            // Clear the transaction from the scope
+            _hub.ConfigureScope(scope => scope.ResetTransaction(this));
+
+            // Client decides whether to discard this transaction based on sampling
+            _hub.CaptureTransaction(new Transaction(this));
         }
 
         /// <inheritdoc />