+++ test/Sentry.AspNetCore.Grpc.Tests/ApiApprovalTests.cs
@@ -1,0 +1,17 @@
+using System.Threading.Tasks;
+using Sentry.Tests;
+using VerifyXunit;
+using Xunit;
+
+namespace Sentry.AspNetCore.Grpc.Tests
+{
+    [UsesVerify]
+    public class ApiApprovalTests
+    {
+        [Fact]
+        public Task Run()
+        {
+            return typeof(SentryGrpcInterceptor).Assembly.CheckApproval();
+        }
+    }
+}