+++ test/Sentry.Tests/ApiApprovalTests.cs
@@ -1,0 +1,16 @@
+using System.Threading.Tasks;
+using VerifyXunit;
+using Xunit;
+
+namespace Sentry.Tests
+{
+    [UsesVerify]
+    public class ApiApprovalTests
+    {
+        [Fact]
+        public Task Run()
+        {
+            return typeof(SentrySdk).Assembly.CheckApproval();
+        }
+    }
+}